[export] [CallTracer] report transport byte counts directly to `CallTracer`
Previously, on the client side, we passed the transport byte counts to the
`CallTracer` as part of the `RecordReceivedTrailingMetadata()` method.
This made sense in the legacy filter stack, because the transport
returns these stats as part of the `recv_trailing_metadata` op.
However, in the new call v3 stack, these stats are not part of receiving
server trailing metadata; instead, they are available only at call
finalization time.

While thinking about a way to address that problem, we realized that we
can avoid some unnecessary per-call memory usage by having the transport
directly report the byte counts to the `CallTracer` as they are read or
written, and have the `CallTracer` aggregate those counts if it needs
to report that information.  This will ultimately get us to a place
where we don't use memory for these stats at all unless they are
actually being used.  (In the short term, it actually increases memory
usage, but we can separately do some work to avoid the memory usage in
the transport by removing the `grpc_transport_stream_stats` struct from
the legacy filter API.)

----
DO NOT SUBMIT. This PR is for testing purposes only. [cl/646264604](http://cl/646264604)

PiperOrigin-RevId: 646264604